fix: Handle absence of deleted tickets in mixin

+++ app/mixins/event-wizard.js
@@ -233,6 +233,8 @@ export default Mixin.create(MutableArray, CustomFormMixin, {
 });
 
 function destroyDeletedTickets(deletedTickets) {
+  if (!deletedTickets)
+    return // This mixin may be used in other steps not containing tickets
   deletedTickets.forEach(ticket => {
     ticket.destroyRecord();
   });
